+++ src/shared/utils/index.ts
@@ -1,7 +1,14 @@
+import { parseBoolean } from './formDataUtils';
 export { computeExpirationDays, formatDateTime } from './dateUtils';
 export type { FormatDateTimeOptions } from './dateUtils';
 
-export { downloadFile, formatFileSize, isViewableFileType, parseFileSize } from './fileUtils';
+export {
+	downloadFile,
+	formatFileSize,
+	isViewableFileType,
+	parseFileSize,
+	exportToCsv,
+} from './fileUtils';
 export type { FormatFileSizeOptions } from './fileUtils';
 
 export { convertTransparencyToHex, sortFields, splitName } from './stringUtils';
@@ -21,4 +28,4 @@ export type { ValidationRule } from './validators';
 
 export { buildLinkUrl } from './urlBuilder';
 
-export { exportToCsv } from './exportToCsv';
+export { parseBoolean, emptyToNull } from './formDataUtils';
